+++ R/mesh.R
@@ -88,6 +88,14 @@ make_mesh <- function(data, xy_cols,
     cli_abort(msg)
   }
 
+  all_x_non_na <- sum(is.na(data[[xy_cols[[1]]]])) == 0L
+  all_y_non_na <- sum(is.na(data[[xy_cols[[2]]]])) == 0L
+  if (!all_x_non_na || !all_y_non_na) {
+    msg <- c("Some coordinates in `xy_cols` were NA.", "
+      Remove or fix these rows before proceeding.")
+    cli_abort(msg)
+  }
+
   if (max(data[[xy_cols[1]]]) > 1e4 || max(data[[xy_cols[2]]] > 1e4)) {
     msg <- paste0(
       "The x or y column values are fairly large. ",
@@ -216,30 +224,11 @@ binary_search_knots <- function(loc_xy,
 #' @param ... Passed to [graphics::plot()].
 #'
 #' @importFrom graphics points
-#' @return `plot.sdmTMBmesh()`: A plot of the mesh and data points. If
-#'   \pkg{ggplot2} is installed, a \pkg{ggplot2} object is
-#'   returned, otherwise a base graphics R plot is returned. To make your own,
-#'   pass `your_mesh$mesh` to `inlabru::gg()`.
+#' @return `plot.sdmTMBmesh()`: A plot of the mesh and data points. To make your
+#'   own \pkg{ggplot2} version, pass `your_mesh$mesh` to `inlabru::gg()`.
 #' @rdname make_mesh
 #' @export
 plot.sdmTMBmesh <- function(x, ...) {
-  # r1 <- requireNamespace("inlabru", quietly = TRUE)
-  # r2 <- requireNamespace("ggplot2", quietly = TRUE)
-  # if (r1 && r2) {
-  #   dat <- data.frame(
-  #     x = x$loc_xy[,1,drop=TRUE],
-  #     y = x$loc_xy[,2,drop=TRUE]
-  #   )
-  #   ggplot2::ggplot() +
-  #     # inlabru::gg(x$mesh, ext.color = "grey20", ext.linewidth = 0.5, edge.color = "grey50") +
-  #     ggplot2::coord_sf() +
-  #     fmesher::geom_fm(data = x$mesh) +
-  #     ggplot2::geom_point(
-  #       data = dat,
-  #       mapping = ggplot2::aes(x = .data$x, y = .data$y), alpha = 0.4, pch = 20, colour = "#3182BD") +
-  #     # ggplot2::coord_fixed() +
-  #     ggplot2::labs(x = x$xy_cols[[1]], y = x$xy_cols[[2]])
-  # } else {
     plot(x$mesh, main = NA, edge.color = "grey60", asp = 1, ...)
     points(x$loc_xy, pch = 21, cex = 0.3, col = "#00000080")
     points(x$loc_centers, pch = 20, col = "red")